requirements.txt
python-dotenv==0.21.0
watchdog~=2.2.1